commiting and pushing before i break absolutely everything because I'm about to go in a completely different direction
